fix: another unlink bug check exist before unlink

+++ app/Http/Controllers/Backend/AdminUserController.php
@@ -18,7 +18,7 @@ class AdminUserController extends Controller
     	$adminuser = Admin::where('type',2)->latest()->get();
     	return view('backend.role.admin_role_all',compact('adminuser'));
 
-    } // end method 
+    } // end method
 
 
     public function AddAdminRole(){
@@ -28,7 +28,7 @@ class AdminUserController extends Controller
 
 
  public function StoreAdminRole(Request $request){
-   	 
+
 
     	$image = $request->file('profile_photo_path');
     	$name_gen = hexdec(uniqid()).'.'.$image->getClientOriginalExtension();
@@ -60,7 +60,7 @@ class AdminUserController extends Controller
 		'type' => 2,
 		'profile_photo_path' => $save_url,
 		'created_at' => Carbon::now(),
-		 
+
 
     	]);
 
@@ -71,7 +71,7 @@ class AdminUserController extends Controller
 
 		return redirect()->route('all.admin.user')->with($notification);
 
-    } // end method 
+    } // end method
 
 
 
@@ -80,19 +80,22 @@ class AdminUserController extends Controller
     	$adminuser = Admin::findOrFail($id);
     	return view('backend.role.admin_role_edit',compact('adminuser'));
 
-    } // end method 
+    } // end method
 
 
 
 
  public function UpdateAdminRole(Request $request){
-    	
+
     	$admin_id = $request->id;
     	$old_img = $request->old_image;
 
     	if ($request->file('profile_photo_path')) {
 
-    	unlink($old_img);
+            if (file_exists($old_img)) {
+                unlink($old_img);
+            }
+
     	$image = $request->file('profile_photo_path');
     	$name_gen = hexdec(uniqid()).'.'.$image->getClientOriginalExtension();
     	Image::make($image)->resize(225,225)->save('upload/admin_images/'.$name_gen);
@@ -101,7 +104,7 @@ class AdminUserController extends Controller
 	Admin::findOrFail($admin_id)->update([
 		'name' => $request->name,
 		'email' => $request->email,
-		 
+
 		'phone' => $request->phone,
 		'brand' => $request->brand,
 		'category' => $request->category,
@@ -138,7 +141,7 @@ class AdminUserController extends Controller
     	Admin::findOrFail($admin_id)->update([
 		'name' => $request->name,
 		'email' => $request->email,
-		 
+
 		'phone' => $request->phone,
 		'brand' => $request->brand,
 		'category' => $request->category,
@@ -158,7 +161,7 @@ class AdminUserController extends Controller
 		'alluser' => $request->alluser,
 		'adminuserrole' => $request->adminuserrole,
 		'type' => 2,
-		 
+
 		'created_at' => Carbon::now(),
 
     	]);
@@ -170,9 +173,9 @@ class AdminUserController extends Controller
 
 		return redirect()->route('all.admin.user')->with($notification);
 
-    	} // end else 
+    	} // end else
 
-    } // end method 
+    } // end method
 
 
  	public function DeleteAdminRole($id){
@@ -190,8 +193,7 @@ class AdminUserController extends Controller
 
 		return redirect()->back()->with($notification);
 
- 	} // end method 
+ 	} // end method
 
 
 }
-
